# swagger.yaml
openapi: 3.0.3
info:
  title: Inventory & Sales Management API
  description: >
    API for managing users, authentication, products, categories, suppliers,
    purchase orders, sales orders, and inventory audit logs.
  version: 1.0.0
servers:
  - url: http://localhost:7070
    description: Local development server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /products:
    get:
      tags:
        - Products
      summary: Fetch catalog products
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: Create a product line
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Role:
      type: string
      enum: [ADMIN, MANAGER, STAFF]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        price:
          type: number
          format: float
        costPrice:
          type: number
          format: float
        quantityInStock:
          type: integer
        reorderLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [sku, name]
      properties:
        sku:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
        price:
          type: number
        costPrice:
          type: number
        quantityInStock:
          type: integer
        reorderLevel:
          type: integer
...